homework2/requirements.txt
asgiref==3.7.2
Django==4.2.11
djangorestframework==3.16.1
gyp==0.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytz==2024.1
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
supervisor==4.2.5
wheel==0.42.0
